+++ src/nidb/moduleMRIQA.h
@@ -33,7 +33,7 @@ public:
 	~moduleMRIQA();
 
 	int Run();
-	bool QA(int seriesid);
+	bool QA(qint64 seriesid);
 	bool GetQAStats(QString f, double &pvsnr, double &iosnr, QString &msg);
 	bool GetMovementStats(QString f, double &maxrx, double &maxry, double &maxrz, double &maxtx, double &maxty, double &maxtz, double &maxax, double &maxay, double &maxaz, double &minrx, double &minry, double &minrz, double &mintx, double &minty, double &mintz, double &minax, double &minay, double &minaz, QString &msg);
 	void GetMinMax(QVector<double> a, double &min, double &max);
